+++ annotations/lib/mockito_builder_annotations.dart
@@ -1,6 +1,0 @@
-library mockito_builder_annotations;
-
-class GenerateMocker {
-  final List<Type> types;
-  const GenerateMocker(this.types);
-}