+++ lib/mockor.dart
@@ -1,0 +1,3 @@
+library mockor;
+
+export 'src/annotations/generate_mocker.dart';